viraai.io-ai-project/requirements.txt
requests
python-dotenv
fastapi
uvicorn

